refactored some code

+++ src/lobby/Usecase.ts
@@ -1,20 +1,25 @@
-import { Repository as LobbyRepository, Lobby, Presenter } from './'
-import { Repository as PlayerRepository, Player } from '../player'
+import { Repository as LobbyRepository, Lobby, Presenter, Event as LobbyEvent, EventBus as LobbyEventBus } from './'
+import { Player, Usecase as PlayerUsecase } from '../player'
 import { Repository as SessionRepository, Session } from '../session'
 import * as dto from './dto'
-import { Event, EventBus } from './EventBus'
+import { Event as PlayerEvent , EventBus as PlayerEventBus } from '../player'
 
 export class Usecase {
 	private readonly lobbyRepo: LobbyRepository
-	private readonly playerRepo: PlayerRepository
+	private readonly playerUcase: PlayerUsecase
 	private readonly sessionRepo: SessionRepository
-	private readonly evBus: EventBus
+	private readonly lobbyEvBus: LobbyEventBus
+	private readonly playerEvBus: PlayerEventBus
 
-	constructor(lobbyRepo: LobbyRepository, playerRepo: PlayerRepository, sessionRepo: SessionRepository, evBus: EventBus) {
+	constructor(lobbyRepo: LobbyRepository, playerUcase: PlayerUsecase, sessionRepo: SessionRepository, lobbyEvBus: LobbyEventBus, playerEvBus: PlayerEventBus) {
 		this.lobbyRepo = lobbyRepo
-		this.playerRepo = playerRepo
+		this.playerUcase = playerUcase
 		this.sessionRepo = sessionRepo
-		this.evBus = evBus
+		this.lobbyEvBus = lobbyEvBus
+		this.playerEvBus = playerEvBus
+
+		playerEvBus.subscribe(PlayerEvent.PLAYER_CREATED, this.onPlayerConnect.bind(this))
+		playerEvBus.subscribe(PlayerEvent.PLAYER_DISCONNECTED, this.onPlayerDisconnect.bind(this))
 	}
 
 	createLobby(req: dto.CreateLobbyReq): [Session, dto.CreateLobbyRes] {
@@ -23,13 +28,7 @@ export class Usecase {
 		const lobby = new Lobby()
 		this.lobbyRepo.insert(lobby)
 
-		const player = new Player(req.username, lobby)
-		this.playerRepo.insert(player)
-
-		const session = new Session(player)
-		this.sessionRepo.insert(session)
-
-		this.evBus.publish(Event.NEW_PLAYER, lobby)
+		const session = this.playerUcase.create(req.username, lobby)
 		return [session, {
 			id: lobby.getID(),
 		}]
@@ -40,13 +39,7 @@ export class Usecase {
 
 		const lobby = this.lobbyRepo.get(req.id)
 
-		const player = new Player(req.username, lobby)
-		this.playerRepo.insert(player)
-
-		const session = new Session(player)
-		this.sessionRepo.insert(session)
-
-		this.evBus.publish(Event.NEW_PLAYER, lobby)
+		const session = this.playerUcase.create(req.username, lobby)
 		return [session, {
 			id: lobby.getID()
 		}]
@@ -57,15 +50,25 @@ export class Usecase {
 		return Presenter.getPlayerRes(players)
 	}
 
-	disconnect(session: Session) {
-		this.sessionRepo.delete(session.getID())
-		this.playerRepo.delete(session.getPlayer().getID())
+	onPlayerConnect(player: Player) {
+		const lobby = player.getLobby()
+		if (lobby === null) {
+			return
+		}
 
-		const lobby = session.getPlayer().getLobby()
+		this.lobbyEvBus.publish(LobbyEvent.PLAYER_LIST_CHANGE, lobby)
+	}
+
+	onPlayerDisconnect(player: Player) {
+		const lobby = player.getLobby()
+		if (lobby === null) {
+			return
+		}
+
 		if (this.lobbyRepo.playersCount(lobby.getID()) === 0) {
 			this.lobbyRepo.delete(lobby.getID())
 		} else {
-			this.evBus.publish(Event.DISCONNECTED_PLAYER, lobby)
+			this.lobbyEvBus.publish(LobbyEvent.PLAYER_LIST_CHANGE, lobby)
 		}
 	}
 }
